Buffs charge rate of Powerpack
This is to bring the powerpack up to a similar level as the Ammopack for mission longevity. Where the Ammopack holds a staggering amount of ammo for a relative few guns, the powerpack is compatible with a wider variety and now recharges much faster. You can still run out of shots under constant fire just as the ammopack can (and I have on one occasion), but neither should be viewed as superior to the other now.