Introduces T-72 Pointman longevity belt
The T-72 is the next development in ammunition feeding technology, serving as a miniaturized belt unit successor to the T-71 adaptive ammopack. It supports a more varied list of calibers for most of the infantry rifles and submachineguns in service with the TGMC, and also includes compatibility with 16 gauge shells for the SH-15 automatic shotgun, both slugs and flechette. The liberty has also been taken of introducing bulk ammunition boxes for 16 gauge shells. You may find the T-72 in the Boxes category, beneath these new shotgun boxes in your Automated Weapons Rack.