Various buffs to mech weaponry
All mech weaponry (or more accurately from a code standpoint, mech ammo) now has IFF properties, allowing the mech to fire past marines. Certain weapons, such as the mech flamer and rocket launcher, still retain incidental effects that can be hazardous to allies.

Most mech weapons have had their damage increased by between 50% and 300% depending on rate of fire and caliber, as well as increases to sundering. The highest sunder weapon is the 30mm airburst cannon on a direct hit, with a bonus projectile count of 7 and 10 sunder per projectile. It is now a weapon designed for taking on behemoth level threats. The mech rocket now does 100 damage on a direct hit, with 50 penetration, and the impact center now counts as devastation, insta-gibbing anything that happens to be on ground zero and lacks an explosive armor value. The mech flamer now creates a burst of extra hot blue flames from the impact of its fireball, spreading out three tiles from the epicenter and burning for up to one minute. It is designed as a defensive weapon. In addition, a direct impact by the fireball now deals significantly more damage. Finally, as far as the drastic changes go, I've vastly improved the fire rate and accuracy of the vulcan, doubled its ammunition stores, and increased its accurate range from 5 to 7.