Reworks Ship Technician into Ordnance Technician
The Ordnance Technician is chiefly responsible for making sure the OB is loaded to the specifications of the FC and/or SLs. Secondarily, the OT also makes use of the Ordnance Workshop, a small addition to the OB/Railgun room of each shipmap consisting of a table, a chem dispenser, and an Ordnance Workshop vendor, which dispenses the components needed to make small and large chem grenades, as well as the new Satchel Charge, which holds up to four large beakers. This is part one of a larger project. In the next update to this project, I will be adding recipes for various components needed to produce some of the various types of rockets and shells normally ordered from req, as well as a custom landmine assembly.