Reverts a sandbagged change to the SH-35
In the recent 'Great Respritening' PR by Davidththird, without proper documentation, not only took it upon himself to nerf the firedelay of the SH-35 to 1.4 seconds between shots instead of 0.4, but also decided to be snarky about it in the inline comment.

In order to enforce proper coder ettiquette, this salt PR change sandbagged within in a visual rework should be reverted.